requirements.txt
datasets==2.18.0
chromadb==0.4.22
sentence-transformers==2.3.1
google-generativeai==0.3.2
python-dotenv==1.0.0
tqdm==4.66.1